index: change RocksDB key schema for mappings & alignments
These were formerly keyed by node ID & the first 8 characters of the hex SHA1 of the alignment. This was not sufficient to ensure uniqueness for high-depth (birthday problem), leading to a little bit of lost data loading the index.

This diff instead keys by node ID and a unique nonce, with the nonce state persisted in the database.