resolve many problems with simultaneous paired mapping
Smooth alignments which have small indels or multiple mappings in a single
node after alignment normalization.

Use path-based fragment distances in the MEM Markov model transition function
for pairs.

Set the maximum fragment length to 10kb... 100kb is not reasonable.

Remove a lot of code devoted to the old-style paired end mapper. I will bring
this back in with a separate function and a refactor in the next commit.