sort alignments by score, then complexity of alignment description
This is meant to prefer simpler alignments when the score is identical. The
idea is that we might overly-map to regions that are complex in the graph. It's
just a heuristic though to make a decision in a case where we have little
information.

It shaves a few percent off our absolute false negative rate.