straight GSSW: avoid patching MEM-based partial alignments
We were having a lot of problems with the MEM threader due to the complex
patching routine. This steps back from that and implements a direct local
alignment process for each cluster. The same clustering logic of the MEM
threader is used though. Performance appears to be identical with the previous
version.