use the paired traceback mode in simultaneous MEM alignment
The traceback model can be told to handle paired end tracebacks differently, so
as to try multiple pair candidates between neighboring clusters for each
fragment in the pair. That is, rather than excluding a fragment's cluster from
future tracebacks after we've traced back through it, we consider it again in
cases where we would match with a different cluster for the second fragment in
the pair.

This improves sensitivity to alternative mappings for one or the other read in
the pair.

This seems to help us do better with pairs than single ended reads on yeast.
Previously, the benefits were apparently obscured by other bugs.