"combi" alignments for pairs rather than simultaneous
Add another alignment function for pairs which aligns each read independently
with the single ended aligner, then crosses all pairs, scoring them based on
their individual scores plus a bonus for alignments that fit into the fragment
length distribution that we've observed or set.

This alignment function align_paired_multi_combi implements this model. A flag
is added to the Mapper to allow configuring this dynamically but it is not
hooked up to the command line interface.