avoid bug that caused loss in sensitivity at high hit counts
The MEMMarkovModel features a linear vector which is sorted by MEM position in
the read. A bandwidth lets us build the model out bandwidth steps ahead of the
current MEM, looking for potential continuations of the chain. All is well
except when we get large hit counts for the MEMs, in which case we have huge
runs of the same MEM, for which we only consider the first bandwidth MEMs.

This is clearly broken, and I should have caught it earlier. It hurts us when
we want to increase the hit count very high, such as we want to do in large
genomes. bwa mem uses 500 by default, and we can't get there without changes to
the way we do clustering. The next commits will include a banding feature based
on the two dimensions we care about--- distance in the read and distance in the
graph.