use cached access to the graph to save xg index calls
This completes the basic idea from before by implementing the edge_cache and
writing graph traversal functions using these caches.

The subgraph extraction can be more precise than it was in the xg index, and so
we can get less of the graph when we have seeds available to build the
subgraph. However, we don't win much for mate rescue, because there we are
taking a shot in the dark to try to align the mates nearby.