reseed using the longest common prefix of the parent in the suffix tree
If this value gets high, then it implies that there are a number of exact
repeats of the MEM we are constructing. We then use this to threshold
reseeding.

Reseeding when this max_lcp value gets high makes sense, as in these cases a
long MEM may mask a series of shorter exact matches that cluster at the true
mapping location.

In yeast this marginally helps both runtime and ROC. Higher settings
of the reseed threshold may also become possible, but this should be tested in
human.