improve the runtime of the mapper
Estimate the maximum mapping quality from the distribution of max LCPs for each
MEM and the lengths of the MEMs.

Use this to not try to make too many clusters when we have obviously good or
bad reads.

This apparently comes at no cost to accuracy (in yeast).

To test on human...

Alternative improvement to test: only reseed when we see the estimated max
mapping quality below a certain value.