allow setting a bound on the upper and lower number of attempts per read
The number of alignment attempts is now scaled based on the mapping quality
estimate derived from the pattern of LCP and MEM lengths observed in the suffix
tree traversal in MEM finding.

The maximum is obtained when we have a very low mapping quality.

A minimum is set so that we have to try at least some number of attempts, even
when the mapping quality estimate is very high.