Add basic framework for a log miner
We're going to have one script crunch the test logs and make a report.
Another script will upload the report to the Web somewhere and post
about it on the relevant PR or commit.

The upshot of all this is that we will be able to have nice reports
with performance statistics and maybe even plots in the Github
comments for PRs.