make vg map single-ended mapping 20% faster
* Enable non-traceback gssw alignments in single-ended vg map alignment.
This improves alignment performance by reducing memory allocs and maybe by
reducing cache footprint during local alignment, but the balance of effects
hasn't been measured.

* Always score alignments and generate mapping qualities using the score
adjusted by the soft clip bonus. This is a requirement of the use of the
non-traceback alignments as there is no way to determine if we've utilized the
bonus without traceback. The current default result in us needing to do more
tracebacks, which negates the benefit of non-traceback alignments when we only
need to do a few of them.

* Avoid an expensive check for graph acyclicity by providing the `-m` flag.
Alternative pattern --- detect alignment errors and re-try alignment first
checking for graph acyclicity and unrolling if it isn't acyclic.
Even better alternative --- teach gssw to deal with cyclic graphs.