use the GCSA index length to estimate random match chance
This will adjust more smoothly when we have graphs with varying levels of
complexity (e.g. a pan graph and a ref-only graph). Right now they roughly get
the same random chance match, but in effect we should expect more random
matches to the pan graph due to the much higher kmer content.