find sub mems only for singular matches
We should find sub MEMs not for hits that are likely to be filtered out due to
being too numerous, but for those which have a singluar hit.

The mapper is much faster when we only reseed singluar hits over our reseed
length. It appears to be equally as accurate on yeast, but I need to test on
human.

The sub-mem finding code has been restructured so that we don't hit it
repeatedly during the backwards search part of MEM finding. It will be possible
to retain this benefit and also step back to reseeding all MEMs longer than our
threshold.