un-break chunked alignment, improve pair resolution on complex graphs
The traceback function for the chunked aligner was using a broken distance
estimate. It's amazing it worked at all.

The way the minimum distance was calculated forgot the path-relative
orientations of the positions, which could cause failures in pair cluster
generation on non-linear graphs.

Also this tweaks some settings in msga so the default is likely to produce
reasonable results in reasonable time. This is based on testing in the GRCh38
alt sequences in the testing directory.