implement a much better total mapping attempt scaling algorithm
Attempt the max of --try-at-least and the min of --try-up-to and
max_possible_mq / estimated_mq. This lets us drop the maximum number of
multimaps lower and run faster without loss. Also, check if we get at least 30%
of a read (one in a pair or the read in single) worth of MEM hits in the
reference, and if not cap the mq at the estimate.