trust the assembler
Edge overlaps aren't exactly a precise thing. That's ok, we can make a graph
with errors. If we struggle to align the overlaps we may second-guess the
assembler's decision, which in the end breaks up the assemblies we want to
align against. This changes the overlap check to simply verify tha the overlap
is not longer than either node sequence.