trim the graph and use global alignment to generate surjection
This enables us to align through the alleles that we had in the graph, avoiding
soft clipping. The alternative is a complicated process that requires us to
replace the alleles in our alignments to the graph with the reference-relative
alleles for the segments they align to. Trimming and global aligning is much
simpler and appears to be producing the expected results.