Try using GCC 6 on OS X
The major OS X problem we are having with the unit tests is caused
by throwing any exception causing a segfault inside the standard
library. This might go away by upgrading the compiler.

It's a bit hard to debug this since I can't replicate it locally
since I don't have a Homebrew system.