implement an efficient pruning strategy not based on flawed "kpaths"
vg mod -p -l K -e E should remove edges from the graph where a kmer of length K
would cross E bifurcations. This was not apparently the case in the previous
implementation, and as a result the pruner will now operate slightly
differently. It tends to be slightly more aggressive at a given setting of K
and E, but the difference is in general relatively minor. Tests suggest the
pruned indexes preform at least as well, if not better, than the previous
version.