fix a surjection bug with indels and implement a minimum softclip length
This resolves two issues with surjection.

First, I have resolved the obvious problems with surjection around indels that
was caused alignment simplification. Corrections have been made to the
simplification algorithm and to the calls to it to ensure that we do not reduce
the internal deletions in the alignments when we simplify them prior to
extracting the cigar. This would have generated many errors around indels in
surjectjed BAMs.

Second, to avoid a perfusion of short soft clips caused by the surjection
technique I implement a minimum softclip length (4 by default) at and below
which we simply emit softclips as matches (in the 'M' cigar op sense).