Add and test paired surjection
Now you can pass -i for interleaved GAM input to surject.
If so, it will read and surject in pairs, and produce correct crossreferences.

Otherwise, it will not rename the reads (so paired reads become distinct fragments) and not crossreference them.

Also move SAM escaping into the BAM converter to not URL-encode all SAM output.