correct mapping against unfolded graphs + adjustments to match Jordan's MEM finding
The first issue was basically that we were trying to align a read in one
orientation, but after unfolding the graph may no longer have matched the
orientation we chose. Given that there is no obvious way to detect if we're
flipping the expected orientation, we can simply try both orientations whenever
we expect to have to unfold or unroll the graph.

The second issue resolved here is a continuation of my attempt to match
Jordan's MEM finding from mpmap in the mapper. Fingers crossed...