surject rewrite and mapper parameter changes (make Jenkins happy)
In surject we now use the global alignment on the non-softclipped part of the
read. At the end we use patch_alignment to try to handle softclips that might
be introduced due to the end of the read running into non-reference sequences
in the graph.

This is hopefully the last iteration of vg map parameter tweaks. We'll see what
Jenkins and large scale testing say.