make XG paths cost O(length) rather than O(graph) (version 8)
This prevents us from allocating a vector for the whole space of the graph for
every path.

It increments the XG version.

I thought this would help with #1468 and #1322, but it doesn't do much.  We're
still burning space somehow, somewhere during construction with graphs that
have high node IDs. Maybe we need a global offset for the node id?