filter down our high-abundance hits to a limit rather than throwing them all out
If we get more hits than our hit max, filter down the hits to the max of the
number of extra multimaps (--try-up-to) or max multimaps (-M). This ensures we
get a chance to accurately estimate mapping quality.  Do this even in
reseeding. Ignore the mem match count when doing this, as we need to actually
load the matches in order to filter them. This completes updates from a
previous commit.