Put a hard upper limit on the hit count of locate-able MEMs
This limit is hopefully "big enough".

The real solution I think would be to roll this into locate, and have locate
limit the result count itself and return some kind of flag and a truncated
result set if there are too many hits. But I'm going to implement this instead
because it's easier and will get @cmarkello going faster.

Should fix #1508.