don't patch surjected alignments as this could introduce errors
I'm not sure what I was thinking in doing this. This could introduce errors at
some low rate when the alignment against the reference is patched against the
full graph (!), so it's not worth the risk. We can wait on the new surjection
mode to fix what I meant to fix by patching.