improve cluster subgraph extraction, dagify more and resolve some tests
Walk the match in the subgraph extraction in order to use a smaller expansion.
This appears to fix a few lower quality alignments in tests.

The graph unrolling wasn't going far enough to align through short cycles, so
it has been increased.

Tests depending on msga behavior have been adjusted to be more stable.