reduce paths memory footprint
This removes the string based path naming, replacing it with an internal id in
the paths object. It also removes a map that was used to project the Mapping*
into a path name. The result is faster to load and process grapsh with many
paths and I see a measurable decrease in memory usage.