Add region support for GBWT index
We can now pass a --region flag to vg index to tell it to only look at part of
its phasing VCF.

I refactored region parsing to pull it out of XG where it was for bad
historical reasons.

I have also fixed a bug in GBWT indexing where we would crash if the graph
ended in a variant without a following fixed reference node.