Make vg add more robust
I've upped the sizes of things that it is willing to align directly,
on the assumption that the underlying aligner is less likely to
outright crash now. I have also given it the ability to chop its
input graph and its created nodes to a more reasonable 32 bp, which
is a much smaller default relative to the sizes of the things it
s trying to align, so there should be fewer problems caused by the
sequences it is aligning expanding out to node boundaries.

This should fix #1912.