specify the cairo directory when including
On some systems, it seems that we need to specify that cairo is within
include/cairo.

It's not clear to me why, or if we should make this change. It does fix my
build problems though. If testing is fine then we might as well set it up like
this to avoid future problems.