correct mistakes in chunked alignment
We were considering transitions between the same band in the read, which is
just wrong as these connections should be impossible.

Fixed the number of multimaps to attempt in the AlignmentChainModel. This was
set to the number of multimaps to emit, not the number of multiple mappings to
attempt in order to obtain the best alignment and estimate mapping quality.

Added a graphviz view of the AlignmentChainModel.