Remove raptor and dependants from the Brewfile
I think that having these available from Homebrew is causing our Mac Travis tests to fail, because the headers and/or libraries Homebrew are being picked up by the build process instead of the ones we try to provide. This is somehow leading to a header/library mismatch that makes raptor not work.