Find the node that ate the node to merge through
The Tsin 2007 implementations find the lowest ancestor of the node we are
supposed to be merging through, and merge through that instead. They also cut
off the merged prefix of the node's path.

This adopts that behavior, so we don't get upset that we never found the node
at the other end of the back edge on our path.