Update to latest vcflib
This updates to current vcflib, slightly patched to fix some missing return
statements.

This fixes half of #2733 by getting rifd of the `-lvbcflib` in vcflib, because
the new vcflib has a completely new CMake build system.

It can't be instructed to use vg's HTSlib (see
https://github.com/vcflib/vcflib/issues/286), but apparently it uses a similar
enough HTSlib that I can get it to build and run tests.

We can't CI this on vg's CI until the patches are upstreamed in vcflib; I don't
want to point vg at the vgteam fork again.