Stop trying to align very large variants
This fixes #3154 by not aligning alleles for any very large variants, whether
they look like SVs or not. Since, by the time we get to parsing the alleles,
SVs have had their base-level sequences filled in, they don't look like SVs
anymore, and so we were trying to align their sequences.

Now there's a `-l` on the constructor (default 100), and variants with alleles
longer than that will always be flat.