remove the LRU cache of gbwt positions
The blocker here is that we now have nested parallelism. It's not clear
how to catch the correct LRU cache for a given thread id, as due to
nesting multiple threads may have the same id. It's possible to
restructure things to be serial over snarls and parallel inside of them,
in which case this issue would go away, but I am not sure how useful it
would be.  By running more-fully in parallel, using dynamic scheduling,
we are less likely to get good use of the LRUCache. We can simply use
more threads.