Refactor MinimizerMapper
Change MinimizerMapper to use VectorView to reorder the minimizers by score
insread of actually sorting them, so we can also keep them in the original
order.

Also, change from pairs and vectors to std::array for keeping track of paired
reads, to make it easier to loop over the reads in a pair. Also, define some
types for the "indexes" that point to Alignments based on fragment
cluster/read/alignment number, with lookup functions to reduce the amount of
[.first][.first][.second.first] sort of nonsense.