Allow non-exons to have bad chromosomes
Some GFF3 features aren't used by default, e.g. `region` or other non-gene things. However, if their chromosomes aren't in the graph, that causes a program-halting error. This change moves the check for whether the chromosome is valid to after the check for whether the feature is relevant, allowing graceful ignorance when possible. The program will still error in any situation where the chromosome is important.